refactor: split `state.go` into two separate files
