+++ core/state/state.go
@@ -1,11 +1,9 @@
 package state
 
 import (
-	"encoding/binary"
 	"errors"
 	"fmt"
 	"maps"
-	"math"
 	"runtime"
 	"slices"
 	"sync"
@@ -13,7 +11,6 @@ import (
 	"github.com/NethermindEth/juno/core"
 	"github.com/NethermindEth/juno/core/crypto"
 	"github.com/NethermindEth/juno/core/felt"
-	"github.com/NethermindEth/juno/core/trie2"
 	"github.com/NethermindEth/juno/core/trie2/trienode"
 	"github.com/NethermindEth/juno/core/trie2/trieutils"
 	"github.com/NethermindEth/juno/db"
@@ -65,16 +62,6 @@ func New(stateRoot *felt.Felt, db *StateDB, batch db.Batch) (*State, error) {
 		stateObjects: make(map[felt.Felt]*stateObject),
 		batch:        batch,
 	}, nil
-}
-
-// ContractStorage on State checks dirty storage cache before falling back to the trie.
-func (s *State) ContractStorage(addr, key *felt.Felt) (felt.Felt, error) {
-	obj, err := s.getStateObject(addr)
-	if err != nil {
-		return felt.Felt{}, err
-	}
-
-	return obj.getStorage(key)
 }
 
 // Applies a state update to a given state. If any error is encountered, state is not updated.
@@ -589,344 +576,4 @@ func (s *State) compareContracts(a, b felt.Felt) int {
 	}
 
 	return len(contractB.dirtyStorage) - len(contractA.dirtyStorage)
-}
-
-// StateReader is a read-only view of the state at a given root.
-type StateReader struct {
-	initRoot     felt.Felt
-	db           *StateDB
-	contractTrie *trie2.Trie
-	classTrie    *trie2.Trie
-}
-
-// NewStateReader creates a read-only view of the state at the given root.
-// Should be used for read operations that don't require state mutations.
-func NewStateReader(stateRoot *felt.Felt, db *StateDB) (*StateReader, error) {
-	contractTrie, err := db.ContractTrie(stateRoot)
-	if err != nil {
-		return nil, err
-	}
-
-	classTrie, err := db.ClassTrie(stateRoot)
-	if err != nil {
-		return nil, err
-	}
-
-	return &StateReader{
-		initRoot:     *stateRoot,
-		db:           db,
-		contractTrie: contractTrie,
-		classTrie:    classTrie,
-	}, nil
-}
-
-func (s *StateReader) ContractClassHash(addr *felt.Felt) (felt.Felt, error) {
-	contract, err := GetContract(s.db.disk, addr)
-	if err != nil {
-		return felt.Felt{}, err
-	}
-	return contract.ClassHash, nil
-}
-
-func (s *StateReader) ContractNonce(addr *felt.Felt) (felt.Felt, error) {
-	contract, err := GetContract(s.db.disk, addr)
-	if err != nil {
-		return felt.Felt{}, err
-	}
-	return contract.Nonce, nil
-}
-
-func (s *StateReader) ContractStorage(addr, key *felt.Felt) (felt.Felt, error) {
-	storageTrie, err := s.db.ContractStorageTrie(&s.initRoot, addr)
-	if err != nil {
-		return felt.Zero, err
-	}
-
-	path := storageTrie.FeltToPath(key)
-	v, err := trieutils.GetNodeByPath(
-		s.db.disk,
-		db.ContractTrieStorage,
-		(*felt.Address)(addr),
-		&path,
-		true,
-	)
-	if err != nil {
-		if errors.Is(err, db.ErrKeyNotFound) {
-			return felt.Zero, nil
-		}
-		return felt.Zero, err
-	}
-
-	return felt.FromBytes[felt.Felt](v), nil
-}
-
-// ContractStorageLastUpdatedBlock returns the most recent block number at which a given storage
-// slot key of a given contract was last updated.
-func (s *StateReader) ContractStorageLastUpdatedBlock(
-	addr *felt.Address,
-	key *felt.Felt,
-) (uint64, error) {
-	prefix := db.ContractStorageHistoryKey((*felt.Felt)(addr), key)
-	return s.lastUpdatedBlockNumber(prefix, math.MaxUint64)
-}
-
-func (s *StateReader) ContractDeployedAt(addr *felt.Felt, blockNum uint64) (bool, error) {
-	contract, err := GetContract(s.db.disk, addr)
-	if err != nil {
-		if errors.Is(err, ErrContractNotDeployed) {
-			return false, nil
-		}
-		return false, err
-	}
-
-	return contract.DeployedHeight <= blockNum, nil
-}
-
-func (s *StateReader) Class(classHash *felt.Felt) (*core.DeclaredClassDefinition, error) {
-	return GetClass(s.db.disk, classHash)
-}
-
-func (s *StateReader) ClassTrie() (core.Trie, error) {
-	return s.classTrie, nil
-}
-
-func (s *StateReader) ContractTrie() (core.Trie, error) {
-	return s.contractTrie, nil
-}
-
-func (s *StateReader) ContractStorageTrie(addr *felt.Felt) (core.Trie, error) {
-	return s.db.ContractStorageTrie(&s.initRoot, addr)
-}
-
-func (s *StateReader) CompiledClassHash(
-	classHash *felt.SierraClassHash,
-) (felt.CasmClassHash, error) {
-	metadata, err := core.GetClassCasmHashMetadata(s.db.disk, classHash)
-	if err != nil {
-		return felt.CasmClassHash{}, err
-	}
-	return metadata.CasmHash(), nil
-}
-
-func (s *StateReader) CompiledClassHashV2(
-	classHash *felt.SierraClassHash,
-) (felt.CasmClassHash, error) {
-	metadata, err := core.GetClassCasmHashMetadata(s.db.disk, classHash)
-	if err != nil {
-		return felt.CasmClassHash{}, err
-	}
-	return metadata.CasmHashV2(), nil
-}
-
-func (s *StateReader) Commitment(protocolVersion string) (felt.Felt, error) {
-	contractRoot, err := s.contractTrie.Hash()
-	if err != nil {
-		return felt.Felt{}, err
-	}
-	classRoot, err := s.classTrie.Hash()
-	if err != nil {
-		return felt.Felt{}, err
-	}
-	return stateCommitment(&contractRoot, &classRoot, protocolVersion), nil
-}
-
-func (s *StateReader) GetReverseStateDiff(
-	blockNum uint64,
-	diff *core.StateDiff,
-) (core.StateDiff, error) {
-	reverse := core.StateDiff{
-		StorageDiffs:    make(map[felt.Felt]map[felt.Felt]*felt.Felt, len(diff.StorageDiffs)),
-		Nonces:          make(map[felt.Felt]*felt.Felt, len(diff.Nonces)),
-		ReplacedClasses: make(map[felt.Felt]*felt.Felt, len(diff.ReplacedClasses)),
-	}
-
-	for addr, stDiffs := range diff.StorageDiffs {
-		reverse.StorageDiffs[addr] = make(map[felt.Felt]*felt.Felt, len(stDiffs))
-		for key := range stDiffs {
-			value := felt.Zero
-			if blockNum > 0 {
-				oldValue, err := s.ContractStorageAt(&addr, &key, blockNum-1)
-				if err != nil {
-					return core.StateDiff{}, err
-				}
-				value = oldValue
-			}
-			reverse.StorageDiffs[addr][key] = &value
-		}
-	}
-
-	for addr := range diff.Nonces {
-		oldNonce := felt.Zero
-		if blockNum > 0 {
-			var err error
-			oldNonce, err = s.ContractNonceAt(&addr, blockNum-1)
-			if err != nil {
-				return core.StateDiff{}, err
-			}
-		}
-		reverse.Nonces[addr] = &oldNonce
-	}
-
-	for addr := range diff.ReplacedClasses {
-		oldCh := felt.Zero
-		if blockNum > 0 {
-			var err error
-			oldCh, err = s.ContractClassHashAt(&addr, blockNum-1)
-			if err != nil {
-				return core.StateDiff{}, err
-			}
-		}
-		reverse.ReplacedClasses[addr] = &oldCh
-	}
-
-	return reverse, nil
-}
-
-// Returns the storage value of a contract at a given storage key at a given block number.
-func (s *StateReader) ContractStorageAt(addr, key *felt.Felt, blockNum uint64) (felt.Felt, error) {
-	prefix := db.ContractStorageHistoryKey(addr, key)
-	return s.getHistoricalValue(prefix, blockNum)
-}
-
-// Returns the block number at which a given storage slot key of a given contract was
-// last updated, up to and including the given block number.
-func (s *StateReader) ContractStorageLastUpdatedAt(
-	addr *felt.Address,
-	key *felt.Felt,
-	blockNum uint64,
-) (uint64, error) {
-	prefix := db.ContractStorageHistoryKey((*felt.Felt)(addr), key)
-	return s.lastUpdatedBlockNumber(prefix, blockNum)
-}
-
-// Returns the nonce of a contract at a given block number.
-func (s *StateReader) ContractNonceAt(addr *felt.Felt, blockNum uint64) (felt.Felt, error) {
-	prefix := db.ContractNonceHistoryKey(addr)
-	return s.getHistoricalValue(prefix, blockNum)
-}
-
-// Returns the class hash of a contract at a given block number.
-func (s *StateReader) ContractClassHashAt(addr *felt.Felt, blockNum uint64) (felt.Felt, error) {
-	prefix := db.ContractClassHashHistoryKey(addr)
-	return s.getHistoricalValue(prefix, blockNum)
-}
-
-func (s *StateReader) CompiledClassHashAt(
-	classHash *felt.SierraClassHash,
-	blockNumber uint64,
-) (felt.CasmClassHash, error) {
-	metadata, err := core.GetClassCasmHashMetadata(s.db.disk, classHash)
-	if err != nil {
-		return felt.CasmClassHash{}, err
-	}
-	return metadata.CasmHashAt(blockNumber)
-}
-
-func (s *StateReader) getHistoricalValue(prefix []byte, blockNum uint64) (felt.Felt, error) {
-	var ret felt.Felt
-
-	err := s.valueAt(prefix, blockNum, func(val []byte) error {
-		ret.SetBytes(val)
-		return nil
-	})
-	if err != nil {
-		if errors.Is(err, ErrNoHistoryValue) {
-			return felt.Zero, nil
-		}
-		return felt.Zero, err
-	}
-
-	return ret, nil
-}
-
-// Returns the value at the given block number.
-// First, it attempts to seek for the key at the given block number.
-// If the key exists, it means that the key was modified at the given block number.
-// Otherwise, it moves the iterator one step back.
-// If a key-value pair exists, we have found the value.
-func (s *StateReader) valueAt(prefix []byte, blockNum uint64, cb func(val []byte) error) error {
-	it, err := s.db.disk.NewIterator(prefix, true)
-	if err != nil {
-		return err
-	}
-	defer it.Close()
-
-	seekKey := binary.BigEndian.AppendUint64(prefix, blockNum)
-	if !it.Seek(seekKey) {
-		return ErrNoHistoryValue
-	}
-
-	key := it.Key()
-	keyBlockNum := binary.BigEndian.Uint64(key[len(prefix):])
-	if keyBlockNum == blockNum {
-		// Found the value
-		val, err := it.Value()
-		if err != nil {
-			return err
-		}
-
-		return cb(val)
-	}
-
-	// Otherwise, move the iterator backwards
-	if !it.Prev() {
-		// Moving iterator backwards is invalid, this means we were already at the first key
-		// No values will be found beyond the first key
-		return ErrNoHistoryValue
-	}
-
-	val, err := it.Value()
-	if err != nil {
-		return err
-	}
-
-	return cb(val)
-}
-
-// lastUpdatedBlockNumber finds the most recent block number (up to upToBlock) recorded in
-// a history bucket for the given key prefix. All history buckets ([ContractStorageHistory],
-// [ContractNonceHistory], and [ContractClassHashHistory]) share the same key layout:
-// prefix + uint64 block number in big-endian.
-func (s *StateReader) lastUpdatedBlockNumber(
-	historyKeyPrefix []byte,
-	upToBlock uint64,
-) (uint64, error) {
-	it, err := s.db.disk.NewIterator(historyKeyPrefix, true)
-	if err != nil {
-		return 0, err
-	}
-	defer it.Close()
-
-	seekKey := binary.BigEndian.AppendUint64(historyKeyPrefix, upToBlock)
-
-	if it.Seek(seekKey) {
-		seekedKey := it.Key()
-		seekedBlock := binary.BigEndian.Uint64(seekedKey[len(historyKeyPrefix):])
-		if seekedBlock == upToBlock {
-			return upToBlock, nil
-		}
-	}
-
-	if !it.Prev() {
-		return 0, nil
-	}
-
-	foundKey := it.Key()
-	blockNum := binary.BigEndian.Uint64(foundKey[len(historyKeyPrefix):])
-	return blockNum, nil
-}
-
-// Calculate the commitment of the state.
-// Since v0.14.0, the Poseidon hash is always applied even when classRoot is zero.
-func stateCommitment(contractRoot, classRoot *felt.Felt, protocolVersion string) felt.Felt {
-	if classRoot.IsZero() && contractRoot.IsZero() {
-		return felt.Zero
-	}
-
-	ver, _ := core.ParseBlockVersion(protocolVersion)
-	if classRoot.IsZero() && ver.LessThan(core.Ver0_14_0) {
-		return *contractRoot
-	}
-
-	return crypto.PoseidonArray(stateVersion0, contractRoot, classRoot)
 }
